+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased] - ???
+## [v3.5.2] - 2022-12-04
 ### New
 - Arduino: Add "version" console command
 - NINA/AirLift: Add command to disconnect gamepad ([Github Issue #26][github_issue_26])